requirements.txt
google-genai
httpx
ollama
neo4j